pyproject.toml
[project]
name = "sadakai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "tiktoken>=0.12.0",
]
